refactor(logs): Split logs handling in a separated file
